Time: 3 ms (100%), Space: 43.8 MB (46.35%) - LeetHub

+++ 0201-bitwise-and-of-numbers-range/0201-bitwise-and-of-numbers-range.java
@@ -1,6 +1,6 @@
 class Solution {
     public int rangeBitwiseAnd(int left, int right) {
-        short i=0;
+        int i=0;
         while(left != right){
             left >>= 1;
             right >>= 1;
